refactor: delete user confirmation dialog
